clean up
move modules and classes out of base
